hospital-be/pyproject.toml
[project]
name = "hospital"
version = "0.1.0"
description = ""
authors = [
    {name = "Avanish Gupta",email = "avanishcodes+github@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "psycopg-binary (>=3.2.9,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]

[tool.poetry]
package-mode=false

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
alembic = "^1.16.4"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
